feat(perf): Favor a single pipeline config lookup
Fetching a single pipeline config _should_ be more efficient than
fetching all pipelines (for an application) and filtering in
`orca`.

This is somewhat nuanced as the "all pipelines for application" is
likely being served out of the in-memory cache.

We have seen what looks like serialization-related load and this PR
aims to offer some relief.