chore(build): Update cloudbuild.yaml file
The currently checked-in Google Cloud Build configs have not
been maintained and are no longer used when producing the official
Spinnaker releases. Instead, the build process generates a GCB
config on-the-fly and uses that.

To make the process more clear and make it easier for others to
reproduce the builds, commit the GCB config that is auto-generated
to the repo. Also, delete the unused cloudbuild-tagged.yaml.