feat(plugins): Added ability to create plugin stages
 ## Motivation

This adds the ability to add very simple Stages as plugins. These Stages are very limited in scope. This allows people to create custom Spinnaker stages without having to dig into and understand all of the Orca codebase. This also means that stages can be added without needing to create a custom version or wait till pull requests are merged.

 ## How to use

To build a plugin stage, one needs to implement the `Stage` interface. The `getName` method is used to determine the Stage's name. The `execute` method takes in a `StageInput<YourClassHere>`. `YourClassHere` is a class you create that has the definition of everything your `Stage` needs to run. At runtime, every `Stage` will be registered with the `StageResolver` after being turned into an `ApiStageDefinitionBuilder`.

The `ApiStageDefinitionBuilder` will build a `taskGraph` with a single `ApiTask` which introspects the type required by the `execute` method. The pipeline context will be converted to your `YourClassHere` and then will call the `execute` method, passing in the correct information.

A quick example on how to implement the Orca plugin stage API:
MyStageContext.java
```
    public class MyStageContext {
        public String artifactName;
        public String artifactLocation;
    }
```

MyStage.java
```
    @Component
    public class MyStage implements Stage<MyStageContext> {
        public String getName() { return "myStage"; }

        public <MyStageContext> StageOutput execute(StageInput<MyStageContext> stageInput) {
            StageOutput stageOutput = new StageOutput();
            // TODO add your stage execution stuff here!
            return stageOutput;
        }
    }
```

To see a more complete example check out: https://github.com/armory/stage-plugin

 ## Future

This is the start of the work needed to support plugin stages. As we get more feedback from users of plugins and plugin creators, we will make adjustments to what these stages can do. For example, a future enhancement maybe to add the ability to support `RetryableTask`s.