fix(error handling): add endpoint to mark executions failed
When an execution fails in to start (e.g. before the execution is persisted) in `orca`,
`orca` will mark and store the execution as failed.
Since the artifact resolution is being moved to `echo`, `echo` needs the same ability
to mark executions as failed - otherwise any failures during artifact resolution in
`echo` will turn into silent failures for the user (if no execution is persisted then
the UI can't show any failures associated with that execution).

This change adds an endpoint that `echo` can call to mark the execution failed