feat(orcestration): add support for dynamic task timeout
Currently, task timeouts can't take anything about the execution environment into account (they are essentially a static method).
This change adds the ability for tasks to consider the current `stage` (`execution`) when evaluating timeout.

Additionally, the `onTimeout` method on the task can now return a `TaskResult`.
This allows the tasks to decide if they should be `TERMINAL` or `FAILED_CONTINUE` on timeout and return data to be added to the stage in a standard way

Both are needed for monitored deploy - as the timeout and what action to take depend on the monitor configuration.
This will also become useful for `clouddriver` tasks that are retryable (e.g. backed by sagas)