feat(sql): ability to override SqlHealthIndicator by dynamic config
- SqlHealthActivator is designed to halt queue processing when sql
persistence is in use and the connected database is in a read-only
mode. This is desired behavior when running in a single writer
environment, where write availability is impacted by a failover.
This PR adds a dynamic configuration property,
`sql.overrideSpringHealthCheck` that prevents Spring's Health
Indicator going unhealthy when connected to a read-only database,
allowing Orca to continue serving read-requests for task and pipeline
executions.