fix(source resolver): prefer older servergroup when copying capacity
If there are multiple possible server groups to choose as the source,
and all are of the same size, prefer the older server group.

This supports a use case where multiple simultaneous clone operations
are performed against a single source (the intention is to have
multiple active server groups in the region). Because a server group
can have its capacity manipulated during the deploy lifecycle, we
don't want the second deploy to choose the first deploy as its source
and result in snapshotting a capacity that may have had the min
size pinned high