fix(kayenta): Make sure we destroy canary clusters
This is a follow up to https://github.com/spinnaker/orca/pull/3259 because I rebased my local changes incorrectly.

Sometimes, `DisableClusterStage` can fail. We don't want failures in `DisableClusterStage`
preventing the subsequent `ShrinkCluster` from trying to run and clean up the canary+baseline clusters

This change buids on top of the functionality introduced in https://github.com/spinnaker/orca/pull/3252
to allow sibling stages to run even if a stage fails (correct error code will be propagated to the execution if this happens so it's clear to the user)