fix(notifications+spel): follow up to #3311
This change makes the app notifications consistent with pipeline notifications in terms of how they are processed with SpEL
Related change: https://github.com/spinnaker/orca/pull/3311

The context wasn't being constructed properly so expression like `{execution.id}` would work in notification defined on a pipeline but not on the app.
This was because the context was `{execution: {id: ...,}}` vs. `{id: ...,}`.
This change makes the two consistent (and also consisten with all other places we use spel where context = `{execution: ..., trigger: ...,}`