fix(spel): make contextParameterProcess.process side-effect free
With SpEL v4 Eval Vars calls process once for each variable, that doesn't work because .process has sideeffects and mutates context in a way that prevents it from running twice (bad, bad).
The culprit is `precomputeValues` which expands and does naughty things with the trigger. Making a copy of the context is a simple way to prevent the sideeffects