fix(monitored deploy): Give more leniency to scale down operations
During a rolling deploy, if `scaleDown`, `notifyDeployCompleted`, or `shrinkCluster` fail and auto rollback was requested, a rollback will be performed.
This is suboptimal because you can easily argue that these operations are non-critical. In fact at this point the deploy is completed (and the old ASG is disabled).
So these are purely cleanup actions and having a perfectly good deploy fail and rollback due to a cleanup action failing is silly.

This change makes the three stages return `FAILED_CONTINUE` instead `TERMINAL` on failure thus not triggering a rollback.
The pipeline will still be marked as `FAILED_CONTINUE`