refactor(artifacts): Rename ArtifactResolver to ArtifactUtils
This class does a lot more than resolve artifacts, and an upcoming
commit will pull out the resolving-specific functionality into its
own, smaller class.  In order to leave the ArtifactResolver class
name available for that new class, rename this to ArtifactUtils.

In general, *Utils classes tend to get a lot of functionality that
should live in better purpose-built classes so are often not a great
idea. In this case, the class already has a bunch of functionality so
the name accurately describes it; the upcoming commit will pull out
one part of the functionality and hopefully future refactors will
further pare down this class.