refactor(artifacts): Add ArtifactResolver class
The ArtifactUtils class handles a lot of functionality and is difficult
to understand. Simplify this by creating an ArtifactResolver class to
handle resolving artifacts, so that functionality can be removed from
ArtifactUtils.

One important change in this class is that it does not mutate any of the
Artifact or ExpectedArtifact instances it receives; it returns the
results of its work (containing new instances) which the caller should
then use as appropriate. This will help reduce bugs where the caller is
depending on a mutation happening several calls below the call stack,
and brings us closer to making Artifact and ExpectedArtifact immutable.