refactor(artifacts): Update callers to use ArtifactResolver
This commit updates all artifact resolving code to use the new
ArtifactResolver class.

A few things of note in the changes:
* FindArtifactFromExecutionTaskTest was mocking resolveExpectedArtifacts
to return resolvedArtifacts; this was subtly different than the actual
function because the actual resolvedArtifacts function mutated the
input expectedArtifacts to add a boundArtifact to each. As the
resolveExpectedArtifacts method has been moved to a simpler class that
we don't need to mock, the test is calling the actual method now and
so we need to update the expected result to account for the fact that
the resolvedExpectedArtifacts now have a bound artifact.
* All of the tests deleted from ArtifactUtilsSpec have equivalent new
tests in ArtifactResolverTest which test the new implementation of
artifact resolving.
* The one changed test in ArtifactUtilsSpec (...sets the bound
artifact...) is a result of fixing what was a bug in the prior code.
As the prior code was mutating each ExpectedArtifact, we were mutating
the expected artifacts on the pipeline itself (and testing that).
resolveArtifacts is storing the resolved artifacts in the trigger,
under resolvedArtifacts; this is where downstream stages look to find
bound artifacts and where we should be testing that the bound artifact
is set. The test now checks that the bound artifacts are set in the
trigger.