perf(artifacts): Lazily compute past pipeline artifacts
The resolveArtifacts function always passes the prior artifacts
to the ArtifactResolver (except when it short-circuits in the
event of no expected artifacts). Looking up artifacts from prior
executions can be expensive, so ideally we'd defer looking these
up until they are actually needed.

This change updates the constructor for ArtifactResolver to accept
a Supplier of prior artifacts; it will memoize this supplier and
only invoke it if the resolution requires looking up prior artifacts
(that is, the first time it encounters an expected artifact with
usePriorAritfact set to true that does not match one of the current
artifacts).

This will also allow us in the next commit to remove the
short-circuiting for the case where there are no expected artifacts
without worrying about negatively affecting performance.