refactor(artifacts): A few minor refactors of resolveArtifacts
Replace the List<ExpectedArtifact> with an ImmutableList<> as we
never modify it. (And there's a tiny performance benefit that
ArtifactResolver.getInstance should short-circuit the copy.)

Replace .distinct() collecting to a List<> of receivedArtifacts
with directly collecting to toImmutableSet() as it will remove
duplicates and return a set whose iteration order is the encounter
order in the stream.

Remove the short-circuit that returns early if there are no expected
artifacts; the ArtifactResolver is a no-op that returns empty
lists for resolvedArtifacts and resolvedExpectedArtifacts in that
case, so there is no need to special-case it.

Replace the LinkedHashSet code that creates a deduplicated list of
artifacts in the supplied order with an ImmutableSet that does the same.