refactor(artifacts): Make static factory parameters non-null
We never actually need currentArtifacts to be null, so remove the
nullable annotation. To handle priorArtifacts, add an overload that
omits the parameter.

Also update the resolveExpectedArtifacts parameter to be non-null.
In order to ensure a safe non-null chain of custody, also added
Nonnull to the artifacts field in FindArtifactFromExecutionContext,
and while there made it immutable.  (Also slightly changed the
constructor as prior to this the case where both expectedArtifact
and expectedArtifacts were null create a singleton array with a null
element that was likely to create downstream hard-to-debug issues.)