refactor(artifacts): Simplify error handling of incorrect StageContext
Two functions log a warning and return a default value if the input
StageContext is not of the correct type. Instead of putting the
entire (or bulk of the) function in an if/else block, just check
the desired condition and return early if it fails, pulling the rest
of the function out of the if/else block.

In the case of getBoundArtifactForId, just return null early rather
than process an empty list that will eventually return null when
nothing matches.