refactor(artifacts): Add Nonnull and Nullable annotations
ArtifactUtils is generally good about annotating parameters and return
values with Nonnull or Nullable. Add NonnullByDefault to the class so
that all methods, parameters, and fields are non-null by default, and
explicitly annotate those that are nullable. (Also change a nonnull
Boolean to a boolean).

Likewise, add NonnullByDefault to ArtifactResolver and remove all of
the explicit Nonnull annotations.  (Nothing in that class can be null.)