refactor(artifacts): Simplify getExecutionForPipelineId
Let's get out of RxJava Observable land sooner, returning all found
executions as a list, and finding the max in the stream. This has two
advantages:
(1) We can similify start_time_or_id to be a Comparator instead of a
Function<?,?,?>, and use the Comparator class to generate it.
(2) (minor) We can just get the .max() of the stream instead of sorting
the whole thing and getting the first element

Also change getExecutionForPipelineId to return an Optional instead
of being nullable; this makes the calling code a lot cleaner as well.