refactor(artifacts): Simplify getAllArtifacts
The overload of getAllArtifacts that accepts a filter is only ever
called from within the class; make it private.

It's also overkill to make the parameter be an Optional<Predicate>
as we have another overload of the function that doesn't require
the parameter at all; if you're calling this one, you should pass
in an actual predicate.  This simplifes the callers by unwrapping
some Optionals.

As the includeTrigger boolean is always true; remove the
parameter.

Finally, remove the objectMapper.convertValue() converting
the result of trigger.getArtifacts() to List<Artifact>; this
is already the type of the return value of that function.