refactor(artifacts): ArtifactResolver should accept a Java 8 supplier
ArtifactResolver was accepting a guava Supplier (as it is using
some guava utils to create the memoized supplier). Make things
easier on clients by accepting a Java 8 supplier, and converting
it to a Guava supplier within the class, hiding this implementation
detail from clients.

None of the callers actually need to change because they are all
passing in a lambda which satisfies both functional interfaces,
but if at some point in the future a caller has a Supplier<> in
a variable, I'd rather it be able to use it directly instead of
converting to a guava supplier. (A caller that has a guava supplier
will still be able to use it as the guava Supplier interface
extends the java 8 Supplier interface.)