fix(execution): Fix START_TIME_OR_ID comparator
I noticed that the START_TIME_OR_ID comparator violates the general
contract of Comparator in the case where one of hte executions has a
null start time. I haven't noticed this causing actual bugs, but I
think the failure mode would be intermittent cases of
"Comparison method violates its general contract" being thrown
when trying to sort based on this comparator, which would be annoying
and hard to debug, so fixing this now.