fix(execution): Fix double-null case in sorting
The prior commit did not properly handle the case where we compare
two executions with null startTime; instead of calling them equal,
we should compare by id.

This was caught by a test in orca-redis; while it was broken before
this change I guess the test was working even in that broken state.