fix(execution): Fix START_TIME_OR_ID comparator (#3392)
* test(execution): Add tests to START_TIME_OR_ID

The next commit will fix a bug in the START_TIME_OR_ID comparator;
add tests to document the current behavior and ensure that the
bug fix only affects the intended case.

* fix(execution): Fix START_TIME_OR_ID comparator

I noticed that the START_TIME_OR_ID comparator violates the general
contract of Comparator in the case where one of hte executions has a
null start time. I haven't noticed this causing actual bugs, but I
think the failure mode would be intermittent cases of
"Comparison method violates its general contract" being thrown
when trying to sort based on this comparator, which would be annoying
and hard to debug, so fixing this now.

* fix(execution): Fix double-null case in sorting

The prior commit did not properly handle the case where we compare
two executions with null startTime; instead of calling them equal,
we should compare by id.

This was caught by a test in orca-redis; while it was broken before
this change I guess the test was working even in that broken state.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>