fix(ci): Properly set artifacts on CI stage output
The presence of a property file should not affect whether or not artifacts should be added to the output. The Travis and Wercker integrations in Igor don't even care about this property, and the Jenkins integration also works even if the property file is null or blank.

This fix makes it possible to use the "Find artifacts from Execution" stage with Travis and Wercker.