test(kubernetes): Add tests to WaitForManifestStable and Manifest
There are some existing tests in groovy on WaitForManifestStableTask
but they are missing some edge cases that will prove important in the
next commit.  Add a new suite of tests in Java; these duplicate the
Groovy tests and also add more test cases. We'll wait to delete the
Groovy tests until after the refactor in an upcoming commit just to
get a bit more validation that it doesn't break anything.

Also to prepare for some upcoming changes, add tests to the Manifest
class to ensure we can correctly deserialize manifests.