fix(kubernetes): Fix and simplify manifest stability logic
The logic in WaitForManifestStable is quite complex; in particular
it separately distinguishes true, false, and null on the manifest's
status field and has different behavior in each case. Likely due to
this complexity, there are a number of bugs that were surface when
writing tests in the last commit.

Fix these bugs, which also allows us to stop treating a null status
as different from false; the cases where we did were just what was
leading to the bugs noted earlier.

The tests should clearly show the changes, but in particular:
* The status is fixed in a couple of cases where we either failed
too early when one manifest was failed, or eventually returned success
incorrectly even though a manifest had failed on an earlier poll.
* Duplicate log messages are removed; if a manifest has failed, we
now just log that it failed instead of also logging that we are
waiting for it to stabilize.