feat(codebuild): Add more options to AWS CodeBuild stage
1. Add source override. Users could override their source by specifying
Spinnaker artifacts. Only S3 objects, GitHub repo and BitBucket repo are
accepted as valid input. Note that multiple sources are not yet supported,
this will be added as secondarySourceOverride in a later version.

2. Add source version. Users could specify the source version of their
sources. This source version takes precedence of the version specified
in Spinnaker artifact.

3. Add environment variables. Users could specify env vars that would be
used during the build.