chore(interlink): rework MessageFlagger interface
The `isFlagged` method used to return a boolean but would also unintuitively add the event to the queue.

Let's go instead to a more explicit `process` method that throws a new exception type when a message is flagged. It is really meant to be an exception condition, and it gives us an opportunity to provide a richer signal in the exception message than just a fail/pass boolean.