fix(interlink): apply events to a queue+repo operator
For some actions (like cancel, resume, pause), it is not sufficient to invoke the correspond execution repository call, we also need to create some events on the message queue (e.g. to actually cancel running tasks).

The task controller already has the logic to update both the queue and the execution repo, so I'm lifting some of it in a common but somewhat awkwardly named CompoundExecutionOperator. As a result, the same logic will be executed by the task controller and interlink events.