config(core): Push halyard config defaults down to base
This change pushes a number of config settings that were in the
Halyard-specific config file down to the base config so they apply
to all uesrs.

The one potentially breaking change is that echo used to default to
disabled but now defaults to enabled. Given that most users use
Halyard and netflix uses echo, I suspect that nobody was relying on
this to default to off. We can document this as a change in the
release notes (as it's easy to override if desired).