fix(google): Fix race condition in wait for up instances task
There was a failure in the google_regional_server_group test
this morning where the WaitForUpInstancesTask unexpectedly
returned a status of TERMINAL.

After digging into this, it appears that what happenned is
the following:
* The MonitorKatoTask earlier in the pipeline is supposed to wait
for the deploy to finish, and the record certain details about the
deployment in the stage context. In this case, the important field
is the "deploy.server.groups" field, which should contain a list of
server groups deployed.
* In this particular case, the MonitorKatoTask found that the
clouddriver task had completed, but looking in the stage context
it was missing two entries from the task history (including the
final "Orchestration Completed" entry) and had an empty set of
results.
* The WaitForUpInstancesTask immediately returns a TERMINAL status
if tere are no instances to check, which is what caused this failure.

After digging into this, there is actually a race condition in
clouddriver as tasks are not fetched atomically from redis (as I
don't believe it's possible for redis to do this). So if, during
the request to "/tasks/taskId" we first grab the history and results
of the task, then later grab the status, it's possible that we'll have
stale results and history while still showing the status as complete.
(The order is further complicated by the fact that these getters are
actually called by the serialization logic directly, so we don't control
the code where they are called; a task contains a reference to the
redis repository and fetches information on demand in the accessors.)

Likely as a workaround to this, there is a flag katoResultExpected that
cloud providers can set for their deploy stages; in this case, the
MonitorKatoTask will return status RUNNING if the task is completed
but there are no results (and the next poll will likely return the
results).  This flag is set in a number of other places, but not in
GCE deploys. This appears safe as any GCE deploy that doesn't return
a result will fail downstream anyway waiting for up instances. (The
only place we set the flag for GCE is in the CreateServerGroupTask
which always precedes a WaitForUpInstancesTask.)

Also adding a log line to help with any other cases where we
short-circuit WaitForUpInstancesTask due to a lack of instances.