fix(google): Fix CreateServerGroupTaskSpec tests
These tests seem pretty over-specified; ideally we'd check the value
of this key in a single test and have the other tests focus on just
asserting the important keys of the output.  But I'll leave refactoring
these to another day and will just update the relevant GCE field in
each of the tests.