fix(config): Fix accidentally changed bakery config (#3505)
The commit to clean up the config file by pushing some config down
to code accidentally changed the default value of roscoApisEnabled.
This was set as true in orca.yml, but I forgot to flip the default
to true in code when removing the value in orca.yml, which ended up
being a breaking change.

This change flips the default value of the boolean to true in code,
restoring the deafult to what it was prior to the earlier change.