feat(interlink): patch stage support for foreign executions
Some operations like skip wait or manual approvals result in calls to

    PATCH /pipelines/pipelineId/stages/stageId

with a context map values used to augment the designated stage context. The TaskController also updates the stage's `lastModified` field with user and account information, so that manual approval stages can be used to propagate auth info. This is all great as long as we operate on local executions, but for foreign executions (owned by a different orca cluster), we can't just patch read-only stages.

To support this, we:
- make `ExecutionRepository::storeStage` foreign-aware
- for foreign executions, we create a `PatchStageInterlinkEvent` (if supported) that contains the body of the stage
- the stage body includes the updated context map values and `lastModified` information
- we add an `updateStage` method in `CompoundExecutionOperator` and update the `TaskController` to call it
- we also call it from the event's `applyTo` method