fix(peering): Improve peering of deletes
Currently, in order to detect deletes, we need to perform a period FULL diff of the DB. This is EXTREMELY expensive and causes peering delay to be inconsistent.
This change adds what is essentially a log of all "recent" deletes into the new `deleted_executions` table.
The peering agent can then scan that table (remebering the last known "deletion" id) and delete all executions listed in the table.
Deletes older than some time (e.g. 2days) are cleaned up from the table during a delete.

Because execution deletes are now more complicated (need to add a row into the `deleted_executions` table) I had to change how cleanup agents work (in order not to duplicate the deletion code).
While there, I refactored them a little to extract [some] of the common functionality.