fix(plugins): allow more than one SimpleStage impl
Multiple SimpleStages keeps Orca from starting because SimpleTask had a Component annotation and SimpleStage in the constructor. Orca would try to construct all the Tasks for taskResolver and when it got to SimpleTask it wouldn't know which SimpleStage to inject. So SimpleTask obviously shouldn't be a component. But once the Component annotation was removed it would not execute SimpleStages.
The TaskResolver now needs to load all the SimpleStages and wrap each on as a SimpleTask. They need to be keyed by the SimpleStage impl class name, they can't all be keyed as SimpleTask. But you could only use classes of type Task, even though it ultimately just gets turned into a String.
On the other side the RunTaskHandler looked up tasks based on a list of Task components and SimpleTask is no longer in that list. Now RunTaskHandler looks up tasks using the TaskResolver. Doesn't it make more sense to use the TaskResolver to resolve tasks anyway? I also uses the implementingClass to look up the task rather than the RunTask.taskType because it would only be SimpleTask when someone is using SimpleStages. This seems cleaner but we still use RunTask.taskType and resolve the task based on type as a backup mainly for existing tests.