fix(delete): Add index and chunk deletes of old executions
In my prior change (https://github.com/spinnaker/orca/pull/3569) I introduced a concept of a `deleted_executions` table to keep track of ~1day worth of deletes so that peering has much easier time to peer deletes.
This change causes deletion of executiosn to fail sometimes because of a SQL deadlock.

To fix it:
1. Add an index on `deleted_at` - this column is used to delete executions older than `x` time
2. Take the truncation of the `deleted_executions` table out of the big delete transaction