fix(health): OutOfService is a Down health state
When we're waiting for up instances, OutOfService was considered neither Up nor Down, which means that as long as some other health was Up, having one OutOfService would not keep the task waiting.

With this change, we consider OutOfService a Down status, which means that someAreUpAndNoneAreDown will return false, which will keep the task waiting.