fix(gremlin): Fix gremlin config (#3641)
* fix(gremlin): Fix gremlin config

The feature flag for gremlin was removed from deck (along with a
number of other feature flags) in 1.20 but we missed that this
flag was also checked in orca. This change removes checking the
flag in orca (and removes setting the flag in orca.yml).

As orca.yml already sets the base url, we don't need to default the
Value using it (though in the future maybe it's better to push the
defaulting to there rather than have it in the default config).

* fix(gremlin): Push default to where it is used

The prior commit broke a plugin integration test; I think it's because
it somehow isn't loading orca.yml. But either way, let's just push
the default to where it's used anyway rather than have it in orca.yml
far from the relevant code.

(cherry picked from commit c357d423270f82119175c8f661864cb8a0bb0bbf)