fix(titus): Correctly detect image from prior stages
Because `TitusServerGroupCreator` was `@Delegating` work to `AmazonServerGroupCreator` when `AmazonServerGroupCreator::createServerGroupOperation` ran, it thought it was running in "AWS" mode.
The `cloudProvider` variable was set to `aws` (instead of `titus`) and so image lookup from prior stages would always fail to find the titus image to deploy if there was a AWS deploy ancestor stage.

This change ditches the `@Delegate` and uses inheritance as it's more appropriate (and correct) in this case, this way calls to `getCloudProvider` will return the correct cloudprovider

There have been a few attempts at fixing this:
* https://github.com/spinnaker/orca/pull/2960
* https://github.com/spinnaker/orca/pull/2980

Hopefully, this is the last one.